split overlay into repo and handle modules
